requirements.txt
# Core Dependencies
langchain-cerebras>=0.5.0
langchain-core>=0.3.77
langchain>=0.3.27
langgraph>=0.6.8
langgraph-checkpoint>=2.1.1

# Search & AI
exa-py>=1.15.6
cerebras-cloud-sdk>=1.50.1

# Utilities
python-dotenv>=1.1.1
typing-extensions>=4.15.0

# Optional: For Jupyter notebooks
ipykernel>=6.30.1
ipython>=9.6.0

# Optional: For future web UI
# streamlit>=1.28.0
# gradio>=4.0.0
